requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-unfold==0.54.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.9.0
